backend/requirements.txt
# server/requirements.txt
youtube-transcript-api
python-dotenv
openai
requests
openai-whisper
